updated tests for livechat.beforeRoom
